+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * `UploaderComponent::error()` renamed as `getError()`;
 * `Shell::folderChmod()` has a default value for the `$chmod` argument;
 * added `OptionsParser::consume()` method;
+* `TestCaseTrait::assertLogContains()` can now take absolute paths and filename
+    with extension;
 * fixed code and output for the `Shell` class;
 * removed `TestCaseTrait::deleteAllLogs()` method;
 * removed `folderIsWriteable()` global function. Use instead